+++ tests/IPv4HostNormalizerTest.php
@@ -19,7 +19,7 @@ declare(strict_types=1);
 namespace LeagueTest\Uri;
 
 use League\Uri\Components\Host;
-use League\Uri\IPv4Normalizer;
+use League\Uri\IPv4HostNormalizer;
 use League\Uri\Maths\GMPMath;
 use League\Uri\Maths\PHPMath;
 use PHPUnit\Framework\TestCase;
@@ -27,9 +27,9 @@ use function extension_loaded;
 use const PHP_INT_SIZE;
 
 /**
- * @coversDefaultClass \League\Uri\IPv4Normalizer
+ * @coversDefaultClass \League\Uri\IPv4HostNormalizer
  */
-final class IPv4NormalizerTest extends TestCase
+final class IPv4HostNormalizerTest extends TestCase
 {
     /**
      * @dataProvider providerHost
@@ -42,7 +42,7 @@ final class IPv4NormalizerTest extends TestCase
             self::markTestSkipped('The PHP is compile for a x64 OS or loads the GMP extension.');
         }
 
-        self::assertEquals(new Host($expected), IPv4Normalizer::normalize(new Host($input)));
+        self::assertEquals(new Host($expected), IPv4HostNormalizer::normalize(new Host($input)));
     }
 
     /**
@@ -56,7 +56,7 @@ final class IPv4NormalizerTest extends TestCase
             self::markTestSkipped('The GMP extension is needed to execute this test.');
         }
 
-        self::assertEquals(new Host($expected), IPv4Normalizer::normalize(new Host($input), new GMPMath()));
+        self::assertEquals(new Host($expected), IPv4HostNormalizer::normalize(new Host($input), new GMPMath()));
     }
 
     /**
@@ -70,7 +70,7 @@ final class IPv4NormalizerTest extends TestCase
             self::markTestSkipped('The PHP must be compile for a x64 OS.');
         }
 
-        self::assertEquals(new Host($expected), IPv4Normalizer::normalize(new Host($input), new PHPMath()));
+        self::assertEquals(new Host($expected), IPv4HostNormalizer::normalize(new Host($input), new PHPMath()));
     }
 
     public function providerHost(): array